header update
self care is important, logging off for now. Couldn't figure outmany bugs and it's stressing me out. But that's okay, we'll tackle it tomorrow. The most important thing about failure is what you will take away from it. Whether it destroys you or not in that moment, in the morning will it make you say you are a failure or will it make you say you are a student?